Update the note of "Deploy it on HTTP port 9090"
The content of note makes a mistake to beginner. 
If beginner doesn't understand why web application response using same protocol with a protocol of web server 
or overlooks that mechanism. It is difficult to find how to solve it. because it is only error log that 502 gateway error. 
Therefore I think it is needed to shape up note like my revision. Thanks.